PUNTOS 1,2,3,4,9 Y 10 hechos

+++ src/Banco/ClientePremium.java
@@ -3,7 +3,7 @@ package Banco;
 public class ClientePremium extends Cliente {
 
 
-    public ClientePremium(String nombre, String dni, double saldo) {
+    public ClientePremium(String nombre, String dni, float saldo) {
         super(nombre, dni, saldo,true);
 
     }
